pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pixels-lance"
version = "0.1.0"
description = "RPC binary data fetcher and LanceDB storage"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]

dependencies = [
    "lancedb>=0.1.0",
    "requests>=2.28.0",
    "pydantic>=1.10.0",
    "python-dotenv>=0.19.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
pixels-lance = "pixels_lance.cli:main"

[tool.setuptools]
packages = ["pixels_lance"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 100

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
